// src/models/users/dto/oauth2-google.dto.ts
export type OAuth2Google = {
  // iss: 'accounts.google.com',
  iss: string;

  // azp: '753875100965-uqa692t84g20k2m0djocqohr2c4d7a1u.apps.googleusercontent.com',
  azp: string;

  // aud: '753875100965-uqa692t84g20k2m0djocqohr2c4d7a1u.apps.googleusercontent.com',
  aud: string;

  // sub: '116300587110390602634',
  sub: string;

  // email: 'dulh181199@gmail.com',
  email: string;

  // verified: 'true',
  verified: string;

  // at_hash: '19tFDPyaPeF_Kk3P5u0mcw',
  at_hash: string;

  // name: 'Du Lê',
  name: string;

  // picture: 'https://lh3.googleusercontent.com/a-/AOh14Ghot4t_8Ag_mOrvpZiU0BfCPGHq3m-jSzP0iKpGGA=s96-c',
  picture: string;

  // given_name: 'Du',
  given_name: string;

  // family_name: 'Lê',
  family_name: string;

  // locale: 'en',
  locale: string;

  // iat: '1653701516',
  iat: string;

  // exp: '1653705116',
  exp: string;

  // jti: '53e7338f98addeee4794962a58c8cb2652be1ccf',
  jti: string;

  // alg: 'RS256',
  alg: string;

  // kid: '38f3883468fc659abb4475f36313d22585c2d7ca',
  kid: string;

  // typ: 'JWT'
  typ: string;
};